requirements.txt
# testing
pytest
pytest-cov

# common
ndg-httpsclient
pyopenssl
pyasn1